fix and improvement (#79)
+ bug fix: install_fs_filter() 버그
    - 삭제대상 파일이 없는 경우에만 삭제 시도하도록 변경

+ feature: log 모듈 기능 개선
    - `OutputDebugStringA` 함수를 사용해서 디버그 로그를 출력했기때문에 WinDbg 에서 로그 필터링이 곤란했음
    - `DbgPrintEx(DPFLTR_IHVDRIVER_ID)` 로 변경해서 WinDbg 에서도 로그를 확인하기 용이하게 수정함

+ fix: stop_service() 버그
    - 서비스 중지 타임아웃 처리관련 버그 (무한루프/무한 대기) 수정

+++ src/scm_context.cpp
@@ -190,7 +190,7 @@ install_fs_filter(
 	{
 		// 
 		//	StartType = SERVICE_BOOT_START 
-		//	으로 지정하려면 드라이버파일의 경로가 반드시 system32 에 있어야 함
+		//	으로 지정하려면 드라이버파일의 경로가 반드시 system32\drivers 에 있어야 함
 		// 
 	
 		std::wstring windows_dir;
@@ -211,16 +211,19 @@ install_fs_filter(
 		//
 		if (0 != sys_path.str().compare(bin_path))
 		{
-			if (!DeleteFileW(sys_path.str().c_str()))
+			if (is_file_existsW(sys_path.str()))
 			{
-				log_err
-					"Can not delete (old) fs driver. path=%ws, gle=%u",
-					sys_path.str().c_str(),
-					GetLastError()
-					log_end;
-				return false;
+				if (!DeleteFileW(sys_path.str().c_str()))
+				{
+					log_err
+						"Can not delete (old) fs driver. path=%ws, gle=%u",
+						sys_path.str().c_str(),
+						GetLastError()
+						log_end;
+						return false;
+				}
 			}
-
+			
 			if (!CopyFile(bin_path, sys_path.str().c_str(), TRUE))
 			{
 				log_err
@@ -772,7 +775,7 @@ stop_service(
 								&service_status))
 	{
 		log_err
-			"ControlService() failed, service name=%ws, gle = %u",
+			"ControlService() failed, service name=%ws, gle=%u",
 			service_name,
 			GetLastError()
 			log_end;
@@ -784,18 +787,39 @@ stop_service(
 	// 
 	SERVICE_STATUS svc_status = { 0 };
 	uint32_t wait_for_secs = 0;
-	while (QueryServiceStatus(svc_handle.get(), &svc_status))
+	while (true)
 	{
+		if (!QueryServiceStatus(svc_handle.get(), &svc_status))
+		{
+			log_err
+				"QueryServiceStatus() failed. service name=%ws, gle=%u",
+				service_name,
+				GetLastError()
+				log_end;
+			break;
+		}
+
 		if (svc_status.dwCurrentState == SERVICE_STOP_PENDING)
 		{
 			if (wait_for_secs < wait_for_n_secs)
 			{
 				Sleep(1000);
 				wait_for_secs++;
+				continue;
+			}
+			else
+			{
+				//
+				//	timeout
+				//	
+				break;
 			}
 		}
 		else
 		{
+			//
+			//	Ok. Service stopped.
+			//
 			break;
 		}
 	}
